Backend/PythonBackend/requirements.txt
appdirs==1.4.4
astunparse==1.6.3
atomicwrites==1.4.0
backports.weakref==1.0.post1
bidict==0.23.1
blis==0.7.11
catalogue==2.0.10
chart-studio==1.1.0
cloudpathlib==0.16.0
clyent==1.2.2
conda-repo-cli==1.0.75
conda-verify==3.4.2
confection==0.1.4
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
distlib==0.3.8
docstring_parser==0.16
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
et-xmlfile==1.1.0
Flask-SocketIO==5.3.6
fonttools==4.25.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.2
httpcore==1.0.5
huggingface-hub==0.22.2
inflection==0.5.1
jq==1.7.0
jsonpointer==2.1
langchain-google-vertexai==1.0.8
langcodes==3.3.0
munkres==1.1.4
murmurhash==1.0.10
openai==0.28.0
openpyxl==3.0.10
pandas-datareader==0.10.0
patsy==0.5.3
pipenv==2024.0.1
ply==3.11
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==11.0.0
pyasn1-modules==0.2.8
pycurl==7.45.2
PyDispatcher==2.0.5
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
python-engineio==4.9.1
python-lsp-jsonrpc==1.0.0
python-socketio==5.11.2
requests-oauthlib==1.3.0
safetensors==0.4.2
shapely==2.0.5
simple-websocket==1.0.0
spacy==3.7.4
spacy-alignments==0.9.1
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.4
srsly==2.4.8
textblob==0.18.0.post0
thinc==8.2.3
tokenizers==0.15.2
torch==2.2.2
transformers==4.36.2
typer==0.9.4
virtualenv==20.26.3
wasabi==1.1.2
weasel==0.3.4
webencodings==0.5.1
wsproto==1.2.0
